refactor: 이전 방식으로 변경 및 좋아요 관련 비즈니스 로직 FavoriteService로 위치 변경
