+++ board-back/src/main/java/com/zoo/boardback/domain/board/api/BoardController.java
@@ -3,6 +3,7 @@ package com.zoo.boardback.domain.board.api;
 import com.zoo.boardback.domain.auth.details.CustomUserDetails;
 import com.zoo.boardback.domain.board.application.BoardService;
 import com.zoo.boardback.domain.board.dto.request.PostCreateRequestDto;
+import com.zoo.boardback.domain.favorite.application.FavoriteService;
 import com.zoo.boardback.domain.favorite.dto.response.FavoriteListResponseDto;
 import com.zoo.boardback.domain.board.dto.response.PostDetailResponseDto;
 import jakarta.validation.Valid;
@@ -23,6 +24,7 @@ import org.springframework.web.bind.annotation.RestController;
 public class BoardController {
 
   private final BoardService boardService;
+  private final FavoriteService favoriteService;
 
   @PostMapping("")
   public ResponseEntity<Void> createBoard(
@@ -45,7 +47,7 @@ public class BoardController {
   public ResponseEntity<Void> putFavorite(
       @PathVariable int boardNumber,
       @AuthenticationPrincipal CustomUserDetails userDetails) {
-      boardService.putFavorite(boardNumber, userDetails.getUsername());
+      favoriteService.putFavorite(boardNumber, userDetails.getUsername());
       return ResponseEntity.ok().build();
   }
 
@@ -53,6 +55,6 @@ public class BoardController {
   public ResponseEntity<FavoriteListResponseDto> getFavoriteList(
       @PathVariable int boardNumber
   ) {
-    return ResponseEntity.ok(boardService.getFavoriteList(boardNumber));
+    return ResponseEntity.ok(favoriteService.getFavoriteList(boardNumber));
   }
 }
