+++ board-back/src/main/java/com/zoo/boardback/domain/board/application/BoardService.java
@@ -6,11 +6,11 @@ import static java.util.stream.Collectors.toList;
 
 import com.zoo.boardback.domain.board.dao.BoardRepository;
 import com.zoo.boardback.domain.board.dto.request.PostCreateRequestDto;
-import com.zoo.boardback.domain.favorite.dto.response.FavoriteListResponseDto;
 import com.zoo.boardback.domain.board.dto.response.PostDetailResponseDto;
 import com.zoo.boardback.domain.board.entity.Board;
 import com.zoo.boardback.domain.favorite.dao.FavoriteRepository;
 import com.zoo.boardback.domain.favorite.dto.query.FavoriteQueryDto;
+import com.zoo.boardback.domain.favorite.dto.response.FavoriteListResponseDto;
 import com.zoo.boardback.domain.favorite.entity.Favorite;
 import com.zoo.boardback.domain.favorite.entity.primaryKey.FavoritePk;
 import com.zoo.boardback.domain.image.dao.ImageRepository;
@@ -32,7 +32,6 @@ public class BoardService {
   private final BoardRepository boardRepository;
   private final UserRepository userRepository;
   private final ImageRepository imageRepository;
-  private final FavoriteRepository favoriteRepository;
 
   @Transactional
   public void create(PostCreateRequestDto request, String email) {
@@ -65,33 +64,5 @@ public class BoardService {
       boardImageList.add(imageUrl);
     }
     return PostDetailResponseDto.of(board, boardImageList);
-  }
-
-  @Transactional
-  public void putFavorite(int boardNumber, String email) {
-    Board board = boardRepository.findByBoardNumber(boardNumber).orElseThrow(() ->
-        new BusinessException(boardNumber, "boardNumber", BOARD_NOT_FOUND));
-
-    User user = userRepository.findByEmail(email).orElseThrow(() ->
-        new BusinessException(email, "email", USER_NOT_FOUND));
-    FavoritePk favoritePk = new FavoritePk(board, user);
-    Favorite favorite = favoriteRepository.findByFavoritePk(favoritePk);
-    if (favorite == null) {
-      favorite = Favorite.builder()
-          .favoritePk(favoritePk)
-          .build();
-      favoriteRepository.save(favorite);
-      board.increaseFavoriteCount();
-    } else {
-      favoriteRepository.delete(favorite);
-      board.decreaseFavoriteCount();
-    }
-  }
-
-  public FavoriteListResponseDto getFavoriteList(int boardNumber) {
-    Board board = boardRepository.findByBoardNumber(boardNumber).orElseThrow(() ->
-        new BusinessException(boardNumber, "boardNumber", BOARD_NOT_FOUND));
-    List<Favorite> favoriteList = favoriteRepository.findByFavoritePkBoard(board);
-    return FavoriteListResponseDto.from(favoriteList);
   }
 }
