+++ board-back/src/main/java/com/zoo/boardback/domain/favorite/dao/FavoriteRepository.java
@@ -4,9 +4,7 @@ import com.zoo.boardback.domain.board.entity.Board;
 import com.zoo.boardback.domain.favorite.dto.query.FavoriteQueryDto;
 import com.zoo.boardback.domain.favorite.entity.Favorite;
 import com.zoo.boardback.domain.favorite.entity.primaryKey.FavoritePk;
-import com.zoo.boardback.domain.user.entity.User;
 import java.util.List;
-import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
@@ -15,13 +13,9 @@ public interface FavoriteRepository extends JpaRepository<Favorite, FavoritePk> 
 
   Favorite findByFavoritePk(FavoritePk favoritePk);
 
-/*  @Query("SELECT new com.zoo.boardback.domain.favorite.dto.query.FavoriteQueryDto(u.email, u.nickname, u.profileImage) " +
+  @Query("SELECT new com.zoo.boardback.domain.favorite.dto.query.FavoriteQueryDto(u.email, u.nickname, u.profileImage) " +
       "FROM Favorite f " +
       "JOIN f.favoritePk.user u " +
       "WHERE f.favoritePk.board = :board")
-  List<FavoriteQueryDto> findRecommendersByBoard(@Param("board") Board board);*/
-
-  @EntityGraph(attributePaths = "favoritePk.user")
-  List<Favorite> findByFavoritePkBoard(Board board);
-
+  List<FavoriteQueryDto> findRecommendersByBoard(@Param("board") Board board);
 }
