+++ board-back/src/main/java/com/zoo/boardback/domain/favorite/dto/response/FavoriteListResponseDto.java
@@ -4,7 +4,6 @@ import static java.util.stream.Collectors.toList;
 
 import com.zoo.boardback.domain.favorite.dto.object.FavoriteListItem;
 import com.zoo.boardback.domain.favorite.dto.query.FavoriteQueryDto;
-import com.zoo.boardback.domain.favorite.entity.Favorite;
 import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -17,16 +16,16 @@ public class FavoriteListResponseDto {
   private List<FavoriteListItem> favoriteList;
   private boolean isEmpty;
 
-  public static FavoriteListResponseDto from(List<Favorite> favorites) {
+  public static FavoriteListResponseDto from(List<FavoriteQueryDto> favorites) {
     List<FavoriteListItem> favoriteList = null;
     boolean isEmpty = true;
     if (!favorites.isEmpty()) {
       isEmpty = false;
       favoriteList = favorites.stream().map(favorite ->
           FavoriteListItem.builder()
-              .email(favorite.getFavoritePk().getUser().getEmail())
-              .nickname(favorite.getFavoritePk().getUser().getNickname())
-              .profileImage(favorite.getFavoritePk().getUser().getProfileImage())
+              .email(favorite.getEmail())
+              .nickname(favorite.getNickname())
+              .profileImage(favorite.getProfileImage())
               .build()
       ).collect(toList());
     }
